rc50ws/requirements.txt
websockets ~= 10.3
motor ~= 2.5.1
python-dateutil ~= 2.8.2